+++ projects/widget-commands-manager/src/lib/remote-command/remote-command-config.ts
@@ -8,8 +8,10 @@ export class RemoteCommandWidgetConfig {
 
   /** Widget-specific configuration */
   widgetConfig: {
-    /** Device ID of device data model of the Reprieve device */
+    /** Device ID of the Reprieve device from the list of devices (not the device instances) */
     deviceId?: string;
+    /** Device Data Model ID of the Remote Commands data table in the Reprieve Data Model */
+    commandDataModelId?: string;
     /** Device property set ID for command data */
     commandPropertySetId?: string;
     /** Property code key for storing commands in device data */